requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
flask==3.0.0
elasticsearch==8.11.0
urllib3==2.1.0
python-dotenv==1.0.0 